Extract legacy mock hook
