+++ debug/test/browser/devtools/mock-hook.js
@@ -59,3 +59,52 @@ export function parseEmit(args) {
 	}
 	return Array.from(args[1]);
 }
+
+/**
+ * @returns {import('../../../src/internal').LegacyMockHook}
+ */
+export function createLegacyMockHook() {
+	let roots = new Set();
+
+	/** @type {Array<import('../../../src/internal').DevtoolsEvent>} */
+	let events = [];
+
+	function emit(ev, data) {
+		if (ev=='renderer-attached') return;
+		events.push(data);
+	}
+
+	function getFiberRoots() {
+		return roots;
+	}
+
+	function clear() {
+		roots.clear();
+		events.length = 0;
+	}
+
+	function clearEvents() {
+		events.length = 0;
+	}
+
+	let helpers = {};
+
+	return {
+		on() {},
+		inject() { return 'abc'; },
+		onCommitFiberRoot() {},
+		onCommitFiberUnmount(rid, vnode) {
+			if (helpers[rid]!=null) {
+				helpers[rid].handleCommitFiberUnmount(vnode);
+			}
+		},
+		_roots: roots,
+		log: events,
+		_renderers: {},
+		helpers,
+		clear,
+		clearEvents,
+		getFiberRoots,
+		emit
+	};
+}
